+++ app/src/main/java/com/example/maps/MapsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.maps.ui
+package com.example.maps
 
 import androidx.fragment.app.Fragment
